Find the sget binaries in the right directory.

+++ pkgspecs/tools-go-sget.install.sh
@@ -2,4 +2,9 @@
 set -Eeo pipefail
 source /tools/env.sh
 
-cp -v "$YAK_WORKSPACE/workspace/bin/linux_386"/* "$CLFS/tools/i686/bin/"
+if [[ -e "$YAK_WORKSPACE/workspace/bin/linux_386" ]]
+then
+  cp -v "$YAK_WORKSPACE/workspace/bin/linux_386"/* "$CLFS/tools/i686/bin/"
+else
+  cp -v "$YAK_WORKSPACE/workspace/bin"/* "$CLFS/tools/i686/bin/"
+fi
